% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detec3DRelation.R
\name{detecte3Dr}
\alias{detecte3Dr}
\title{A 3D Object relationship detection function}
\usage{
detecte3Dr(tablefrom, tableto, closedfrom, closedto)
}
\arguments{
\item{tablefrom, }{tableto are tables with; closedfrom, closedto}
}
\value{
A table of combined the input tables with result
}
\description{
This function allows you to compare objects saved in two list
}
\keyword{3D}
